requirements.txt
# Streamlit Framework
streamlit==1.31.0

# File Processing Libraries
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# AI & Text Processing
openai==1.12.0
python-dotenv==1.0.1

# Text-to-Speech Libraries
gTTS==2.5.0
pyttsx3==2.90

# Audio Processing
pydub==0.25.1

# Utility Libraries
pathlib==1.0.1
